@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/raft"
-	raftboltdb "github.com/hashicorp/raft-boltdb"
+	raftboltdb "github.com/hashicorp/raft-boltdb/v2"
 	"github.com/hashicorp/serf/serf"
 	"github.com/bodaay/jocko/jocko/fsm"
 	"github.com/bodaay/jocko/jocko/metadata"
@@ -49,7 +49,7 @@ func (b *Broker) setupRaft() (err error) {
 	b.raftTransport = trans
 
 	b.config.RaftConfig.LocalID = raft.ServerID(fmt.Sprintf("%d", b.config.ID))
-	b.config.RaftConfig.StartAsLeader = b.config.StartAsLeader
+	// Note: StartAsLeader was removed from raft library - bootstrap handles this now
 
 	// build an in-memory setup for dev mode, disk-based otherwise.
 	var logStore raft.LogStore
@@ -268,7 +268,7 @@ func (b *Broker) reconcileMember(m serf.Member) error {
 		err = b.handleLeftMember(m)
 	}
 	if err != nil {
-		log.Error.Printf("leader/%d: reconcile member: %s: error: %s", m.Name, b.config.ID, err)
+		log.Error.Printf("leader/%d: reconcile member: %s: error: %s", b.config.ID, m.Name, err)
 	}
 	return nil
 }
@@ -555,7 +555,7 @@ func (b *Broker) handleFailedMember(m serf.Member) error {
 		broker := b.brokerLookup.BrokerByID(raft.ServerID(fmt.Sprintf("%d", n.Node)))
 		if broker == nil {
 			// TODO: this probably shouldn't happen -- likely a root issue to fix
-			log.Error.Printf("trying to assign partitions to unknown broker: %s", n)
+			log.Error.Printf("trying to assign partitions to unknown broker: %v", n)
 			continue
 		}
 		conn, err := defaultDialer.Dial("tcp", broker.BrokerAddr)
