@@ -2,38 +2,51 @@ package log
 
 import (
 	stdlog "log"
-
-	"upspin.io/log"
+	"os"
 )
 
+// Level represents a logging level
+type Level int
+
+const (
+	DebugLevel Level = iota
+	InfoLevel
+	ErrorLevel
+)
+
+var currentLevel = InfoLevel
+
+// Logger interface for logging
+type Logger interface {
+	Printf(format string, v ...interface{})
+	Print(v ...interface{})
+	Println(v ...interface{})
+	Fatal(v ...interface{})
+	Fatalf(format string, v ...interface{})
+}
+
 var (
-	Debug = &logger{l: log.Debug}
-	Info  = &logger{l: log.Info}
-	Error = &logger{l: log.Error}
+	Debug = &logger{level: DebugLevel, l: stdlog.New(os.Stderr, "[DEBUG] ", stdlog.LstdFlags|stdlog.Lshortfile)}
+	Info  = &logger{level: InfoLevel, l: stdlog.New(os.Stderr, "[INFO] ", stdlog.LstdFlags|stdlog.Lshortfile)}
+	Error = &logger{level: ErrorLevel, l: stdlog.New(os.Stderr, "[ERROR] ", stdlog.LstdFlags|stdlog.Lshortfile)}
+	Fatal = &logger{level: ErrorLevel, l: stdlog.New(os.Stderr, "[FATAL] ", stdlog.LstdFlags|stdlog.Lshortfile)}
 )
 
 type logger struct {
 	prefix string
-	l      log.Logger
+	level  Level
+	l      *stdlog.Logger
 }
 
-type Level = log.Level
-
-var (
-	DebugLevel = log.DebugLevel
-	InfoLevel  = log.InfoLevel
-	ErrorLevel = log.ErrorLevel
-)
-
-func New(level log.Level, prefix string) *logger {
-	l := &logger{prefix: prefix}
+func New(level Level, prefix string) *logger {
+	l := &logger{prefix: prefix, level: level}
 	switch level {
-	case log.DebugLevel:
-		l.l = log.Debug
-	case log.InfoLevel:
-		l.l = log.Info
-	case log.ErrorLevel:
-		l.l = log.Error
+	case DebugLevel:
+		l.l = stdlog.New(os.Stderr, "[DEBUG] "+prefix, stdlog.LstdFlags|stdlog.Lshortfile)
+	case InfoLevel:
+		l.l = stdlog.New(os.Stderr, "[INFO] "+prefix, stdlog.LstdFlags|stdlog.Lshortfile)
+	case ErrorLevel:
+		l.l = stdlog.New(os.Stderr, "[ERROR] "+prefix, stdlog.LstdFlags|stdlog.Lshortfile)
 	}
 	return l
 }
@@ -45,14 +58,28 @@ func SetPrefix(prefix string) {
 }
 
 func SetLevel(level string) {
-	log.SetLevel(level)
+	switch level {
+	case "debug":
+		currentLevel = DebugLevel
+	case "info":
+		currentLevel = InfoLevel
+	case "error":
+		currentLevel = ErrorLevel
+	}
 }
 
-func NewStdLogger(l log.Logger) *stdlog.Logger {
-	return log.NewStdLogger(l)
+func NewStdLogger(l Logger) *stdlog.Logger {
+	return stdlog.New(os.Stderr, "", stdlog.LstdFlags)
+}
+
+func (l *logger) shouldLog() bool {
+	return l.level >= currentLevel
 }
 
 func (l *logger) Printf(format string, v ...interface{}) {
+	if !l.shouldLog() {
+		return
+	}
 	if l.prefix == "" {
 		l.l.Printf(format, v...)
 	} else {
@@ -61,6 +88,9 @@ func (l *logger) Printf(format string, v ...interface{}) {
 }
 
 func (l *logger) Print(v ...interface{}) {
+	if !l.shouldLog() {
+		return
+	}
 	if l.prefix == "" {
 		l.l.Print(v...)
 	} else {
@@ -69,6 +99,9 @@ func (l *logger) Print(v ...interface{}) {
 }
 
 func (l *logger) Println(v ...interface{}) {
+	if !l.shouldLog() {
+		return
+	}
 	if l.prefix == "" {
 		l.l.Println(v...)
 	} else {
@@ -92,4 +125,4 @@ func (l *logger) Fatalf(format string, v ...interface{}) {
 	}
 }
 
-var _ log.Logger = (*logger)(nil)
+var _ Logger = (*logger)(nil)
